♻️ fix: refactor Header Ui component
